spark-mcp-tools-db/requirements.txt
fastmcp
uvicorn[standard]
psycopg2-binary
python-dotenv
requests